Refactor TabNet tests by reorganizing and splitting files
The tests for TabNet were restructured to improve maintainability. Tests have been split into separate files based on components such as encoder, decoder, and transformers. Additionally, fixtures were moved to a dedicated `conftest.py` for better modularity.
