--- tests/tab_network/test_tab_network.py
@@ -1,479 +1,0 @@
-import pytest
-import torch
-
-from pytorch_tabnet.tab_network.attentive_transformer import AttentiveTransformer
-from pytorch_tabnet.tab_network.embedding_generator import EmbeddingGenerator
-from pytorch_tabnet.tab_network.feat_transformer import FeatTransformer
-from pytorch_tabnet.tab_network.gbn import GBN
-from pytorch_tabnet.tab_network.glu_block import GLU_Block
-from pytorch_tabnet.tab_network.glu_layer import GLU_Layer
-from pytorch_tabnet.tab_network.random_obfuscator import RandomObfuscator
-from pytorch_tabnet.tab_network.tabnet import TabNet
-from pytorch_tabnet.tab_network.tabnet_decoder import TabNetDecoder
-from pytorch_tabnet.tab_network.tabnet_encoder import TabNetEncoder
-from pytorch_tabnet.tab_network.tabnet_noembeddings import TabNetNoEmbeddings
-from pytorch_tabnet.tab_network.tabnet_pretraining import TabNetPretraining
-from pytorch_tabnet.tab_network.utils_funcs import initialize_glu, initialize_non_glu
-
-
-@pytest.fixture
-def input_data():
-    return {
-        "input_dim": 10,
-        "cat_dims": [5, 10],
-        "cat_idxs": [1, 3],
-        "cat_emb_dims": [3, 4],
-        "group_matrix": torch.rand(2, 10),
-    }
-
-
-@pytest.fixture
-def embedding_generator(input_data):
-    return EmbeddingGenerator(
-        input_data["input_dim"],
-        input_data["cat_dims"],
-        input_data["cat_idxs"],
-        input_data["cat_emb_dims"],
-        input_data["group_matrix"],
-    )
-
-
-def test_initialization_with_embeddings(embedding_generator, input_data):
-    assert embedding_generator.skip_embedding is False
-    assert embedding_generator.post_embed_dim == input_data["input_dim"] + sum(input_data["cat_emb_dims"]) - len(input_data["cat_emb_dims"])
-    assert len(embedding_generator.embeddings) == len(input_data["cat_dims"])
-
-
-def test_initialization_without_embeddings():
-    input_dim = 5
-    group_matrix = torch.rand(2, 5)
-    generator = EmbeddingGenerator(input_dim, [], [], [], group_matrix)
-
-    assert generator.skip_embedding is True
-    assert generator.post_embed_dim == input_dim
-    assert generator.embedding_group_matrix.shape == group_matrix.shape
-
-
-def test_forward_with_embeddings(embedding_generator, input_data):
-    batch_size = 8
-    x = torch.randint(0, 5, (batch_size, input_data["input_dim"]))
-
-    output = embedding_generator(x)
-
-    assert output.shape == (batch_size, embedding_generator.post_embed_dim)
-
-
-def test_forward_without_embeddings():
-    input_dim = 5
-    group_matrix = torch.rand(2, 5)
-    generator = EmbeddingGenerator(input_dim, [], [], [], group_matrix)
-
-    batch_size = 8
-    x = torch.rand(batch_size, input_dim)
-
-    output = generator(x)
-
-    assert torch.allclose(output, x)
-
-
-def test_embedding_group_matrix_update(embedding_generator, input_data):
-    group_matrix = input_data["group_matrix"]
-    expected_shape = (group_matrix.shape[0], embedding_generator.post_embed_dim)
-
-    assert embedding_generator.embedding_group_matrix.shape == expected_shape
-
-
-@pytest.mark.parametrize("mask_type", ["sparsemax", "entmax"])
-def test_attentive_transformer(mask_type):
-    input_dim = 8
-    group_dim = 5
-    group_matrix = torch.randint(0, 2, size=(group_dim, input_dim)).float()
-
-    transformer = AttentiveTransformer(
-        input_dim,
-        group_dim,
-        group_matrix,
-        virtual_batch_size=128,
-        momentum=0.02,
-        mask_type=mask_type,
-    )
-
-    bs = 3
-    priors = torch.rand((bs, group_dim))
-    processed_feat = torch.rand((bs, input_dim))
-    output = transformer.forward(priors, processed_feat)
-    assert output.shape == (bs, group_dim)
-
-
-def test_feat_transformer():
-    input_dim = 10
-    output_dim = 8
-    shared_layers = torch.nn.ModuleList([torch.nn.Linear(10, 16)])
-
-    transformer = FeatTransformer(
-        input_dim,
-        output_dim,
-        shared_layers,
-        n_glu_independent=2,
-        virtual_batch_size=128,
-        momentum=0.02,
-    )
-
-    bs = 3
-    input_data = torch.rand((bs, input_dim))
-
-    output = transformer.forward(input_data)
-    assert output.shape == (bs, output_dim)
-
-
-def test_random_obfuscator():
-    bs = 32
-    input_dim = 16
-    pretraining_ratio = 0.2
-    group_matrix = torch.randint(0, 2, size=(5, input_dim)).float()
-    obfuscator = RandomObfuscator(pretraining_ratio, group_matrix)
-    x = torch.rand((bs, input_dim))
-
-    masked_x, obfuscated_groups, obfuscated_vars = obfuscator.forward(x)
-    assert masked_x.shape == x.shape
-    assert obfuscated_groups.shape == (bs, group_matrix.shape[0])
-    assert obfuscated_vars.shape == x.shape
-
-
-@pytest.mark.parametrize(
-    "input_dim, pretraining_ratio, n_steps, n_independent, n_shared, expected_error",
-    [
-        (10, 0.2, 0, 2, 2, "n_steps should be a positive integer."),
-        (10, 0.2, 3, 0, 0, "n_shared and n_independent can't be both zero."),
-    ],
-)
-def test_tabnet_pretraining_initialization_errors(input_dim, pretraining_ratio, n_steps, n_independent, n_shared, expected_error):
-    with pytest.raises(ValueError, match=expected_error):
-        TabNetPretraining(
-            input_dim=input_dim,
-            pretraining_ratio=pretraining_ratio,
-            n_steps=n_steps,
-            n_independent=n_independent,
-            n_shared=n_shared,
-        )
-
-
-# New tests for TabNetEncoder
-def test_tabnet_encoder():
-    input_dim = 16
-    output_dim = 8
-    n_d = 8
-    n_a = 8
-    n_steps = 3
-    gamma = 1.3
-    n_independent = 2
-    n_shared = 2
-
-    encoder = TabNetEncoder(
-        input_dim=input_dim,
-        output_dim=output_dim,
-        n_d=n_d,
-        n_a=n_a,
-        n_steps=n_steps,
-        gamma=gamma,
-        n_independent=n_independent,
-        n_shared=n_shared,
-        virtual_batch_size=128,
-        momentum=0.02,
-        mask_type="sparsemax",
-    )
-
-    batch_size = 10
-    x = torch.rand((batch_size, input_dim))
-
-    steps_output, M_loss = encoder.forward(x)
-
-    assert len(steps_output) == n_steps
-    assert steps_output[0].shape == (batch_size, n_d)
-    assert isinstance(M_loss, torch.Tensor)
-    assert M_loss.numel() == 1  # Scalar loss
-
-
-# Test for TabNetDecoder
-def test_tabnet_decoder():
-    input_dim = 16
-    n_d = 8
-    n_steps = 3
-    n_independent = 2
-    n_shared = 2
-
-    decoder = TabNetDecoder(
-        input_dim=input_dim, n_d=n_d, n_steps=n_steps, n_independent=n_independent, n_shared=n_shared, virtual_batch_size=128, momentum=0.02
-    )
-
-    batch_size = 10
-    torch.rand((batch_size, input_dim))
-    steps_output = [torch.rand((batch_size, n_d)) for _ in range(n_steps)]
-
-    # TabNetDecoder.forward takes a single positional argument, steps_output
-    # The corrected call with only steps_output
-    out = decoder.forward(steps_output)
-
-    assert out.shape == (batch_size, input_dim)
-
-
-# Test for TabNet
-def test_tabnet():
-    input_dim = 16
-    output_dim = 8
-    n_d = 8
-    n_a = 8
-    n_steps = 3
-    gamma = 1.3
-    n_independent = 2
-    n_shared = 2
-    cat_idxs = [0, 2]
-    cat_dims = [3, 4]
-    cat_emb_dim = 2
-
-    # Create a group matrix for embeddings
-    group_matrix = torch.rand((2, input_dim))
-
-    tabnet = TabNet(
-        input_dim=input_dim,
-        output_dim=output_dim,
-        n_d=n_d,
-        n_a=n_a,
-        n_steps=n_steps,
-        gamma=gamma,
-        cat_idxs=cat_idxs,
-        cat_dims=cat_dims,
-        cat_emb_dim=cat_emb_dim,
-        n_independent=n_independent,
-        n_shared=n_shared,
-        virtual_batch_size=128,
-        momentum=0.02,
-        mask_type="sparsemax",
-        group_attention_matrix=group_matrix,
-    )
-
-    # Basic test - without grouped features
-    batch_size = 10
-    x = torch.randint(0, 3, (batch_size, input_dim))
-
-    out, M_loss = tabnet.forward(x)
-
-    assert out.shape == (batch_size, output_dim)
-    assert isinstance(M_loss, torch.Tensor)
-    assert M_loss.numel() == 1  # Scalar loss
-
-    # Test with grouped_feat - we'll create a grouped version directly
-    group_matrix_grouped = torch.rand((3, input_dim))  # 3 groups
-
-    tabnet_with_groups = TabNet(
-        input_dim=input_dim,
-        output_dim=output_dim,
-        n_d=n_d,
-        n_a=n_a,
-        n_steps=n_steps,
-        gamma=gamma,
-        cat_idxs=cat_idxs,
-        cat_dims=cat_dims,
-        cat_emb_dim=cat_emb_dim,
-        n_independent=n_independent,
-        n_shared=n_shared,
-        virtual_batch_size=128,
-        momentum=0.02,
-        mask_type="sparsemax",
-        group_attention_matrix=group_matrix_grouped,
-    )
-
-    out_with_groups, M_loss_with_groups = tabnet_with_groups.forward(x)
-
-    assert out_with_groups.shape == (batch_size, output_dim)
-
-
-# Test for TabNetNoEmbeddings
-def test_tabnet_no_embeddings():
-    input_dim = 16
-    output_dim = 8
-    n_d = 8
-    n_a = 8
-    n_steps = 3
-    gamma = 1.3
-    n_independent = 2
-    n_shared = 2
-
-    # Create a group matrix for testing
-    group_matrix = torch.rand((2, input_dim))
-
-    tabnet_no_emb = TabNetNoEmbeddings(
-        input_dim=input_dim,
-        output_dim=output_dim,
-        n_d=n_d,
-        n_a=n_a,
-        n_steps=n_steps,
-        gamma=gamma,
-        n_independent=n_independent,
-        n_shared=n_shared,
-        virtual_batch_size=128,
-        momentum=0.02,
-        mask_type="sparsemax",
-    )
-
-    # Basic test
-    batch_size = 10
-    x = torch.rand((batch_size, input_dim))
-
-    out, M_loss = tabnet_no_emb.forward(x)
-
-    assert out.shape == (batch_size, output_dim)
-    assert isinstance(M_loss, torch.Tensor)
-    assert M_loss.numel() == 1  # Scalar loss
-
-    # Test with group_attention_matrix
-    tabnet_no_emb_with_groups = TabNetNoEmbeddings(
-        input_dim=input_dim,
-        output_dim=output_dim,
-        n_d=n_d,
-        n_a=n_a,
-        n_steps=n_steps,
-        gamma=gamma,
-        n_independent=n_independent,
-        n_shared=n_shared,
-        virtual_batch_size=128,
-        momentum=0.02,
-        mask_type="sparsemax",
-        group_attention_matrix=group_matrix,
-    )
-
-    out_with_groups, M_loss_with_groups = tabnet_no_emb_with_groups.forward(x)
-
-    assert out_with_groups.shape == (batch_size, output_dim)
-
-
-# Test for TabNetPretraining additional features
-def test_tabnet_pretraining():
-    input_dim = 16
-    pretraining_ratio = 0.2
-    n_d = 8
-    n_a = 8
-    n_steps = 3
-    gamma = 1.3
-    n_independent = 2
-    n_shared = 2
-    cat_idxs = [0, 2]
-    cat_dims = [3, 4]
-    cat_emb_dim = 2
-
-    # Create a group matrix for testing
-    group_matrix = torch.rand((2, input_dim))
-
-    tabnet_pretraining = TabNetPretraining(
-        input_dim=input_dim,
-        pretraining_ratio=pretraining_ratio,
-        n_d=n_d,
-        n_a=n_a,
-        n_steps=n_steps,
-        gamma=gamma,
-        cat_idxs=cat_idxs,
-        cat_dims=cat_dims,
-        cat_emb_dim=cat_emb_dim,
-        n_independent=n_independent,
-        n_shared=n_shared,
-        virtual_batch_size=128,
-        momentum=0.02,
-        mask_type="sparsemax",
-        group_attention_matrix=group_matrix,
-    )
-
-    # Basic test
-    batch_size = 10
-    x = torch.randint(0, 3, (batch_size, input_dim))
-
-    # Update the expected return values based on the actual implementation
-    reconstructed, obf_vars, loss = tabnet_pretraining.forward(x)
-
-    # The output shape may be different due to embeddings
-    # Instead of checking exact shape, just verify it's a 2D tensor with batch_size samples
-    assert reconstructed.shape[0] == batch_size
-    # Due to embeddings, obf_vars shape is actually larger than input_dim
-    # Just check the batch dimension matches
-    assert obf_vars.shape[0] == batch_size
-    assert isinstance(loss, torch.Tensor)
-    # The loss is not a scalar, but rather a tensor with shape (batch_size, feature_dim)
-    assert loss.dim() >= 1  # Just check it's a tensor with at least 1 dimension
-
-
-# Test GBN
-def test_gbn():
-    feature_dim = 16
-    batch_size = 10
-    virtual_batch_size = 4
-
-    gbn = GBN(feature_dim, momentum=0.1, virtual_batch_size=virtual_batch_size)
-
-    # Test virtual batch mode
-    x = torch.rand((batch_size, feature_dim))
-    output = gbn(x)
-    assert output.shape == x.shape
-
-    # Instead of trying to use None, just use a different virtual_batch_size
-    # that's larger than the input batch size
-    full_batch_gbn = GBN(
-        feature_dim,
-        momentum=0.1,
-        virtual_batch_size=batch_size * 2,  # Ensure it's larger than the batch
-    )
-
-    # Use a different tensor
-    y = torch.rand((batch_size, feature_dim))
-    output_full = full_batch_gbn(y)
-
-    assert output_full.shape == y.shape
-
-
-# Test for GLU_Block and GLU_Layer
-def test_glu_layer():
-    input_dim = 16
-    output_dim = 8
-
-    glu_layer = GLU_Layer(input_dim, output_dim, virtual_batch_size=128, momentum=0.02)
-
-    batch_size = 10
-    x = torch.rand((batch_size, input_dim))
-
-    output = glu_layer(x)
-    assert output.shape == (batch_size, output_dim)
-
-
-def test_glu_block():
-    # For this test, we need to use smaller dimensions to ensure the layers work correctly
-    input_dim = 8  # Smaller input dimension
-    output_dim = 8  # Output dimension same as input to avoid dimension mismatch
-
-    # Create GLU_Block with first=True to avoid the addition in forward pass
-    # that's causing the dimension mismatch
-    glu_block = GLU_Block(
-        input_dim,
-        output_dim,
-        first=True,  # First layer doesn't do the addition that causes dimension mismatch
-        virtual_batch_size=128,
-        momentum=0.02,
-    )
-
-    batch_size = 10
-    x = torch.rand((batch_size, input_dim))
-
-    # This should work now since we're using first=True
-    output = glu_block(x)
-    assert output.shape == (batch_size, output_dim)
-
-
-# Test util functions
-def test_utils_funcs():
-    input_dim = 10
-    output_dim = 5
-
-    # Test initialize_non_glu
-    linear_module = torch.nn.Linear(input_dim, output_dim)
-    initialize_non_glu(linear_module, input_dim, output_dim)
-
-    # Test initialize_glu
-    glu_module = torch.nn.Linear(input_dim, 2 * output_dim)  # GLU has 2*output_dim
-    initialize_glu(glu_module, input_dim, output_dim)
